requirements.txt
selenium==4.10.0
webdriver_manager==3.9.0
jsonschema==4.17.3
Pillow~=9.5.0
requests~=2.31.0